refactor: remove unuse code
