+++ packages/maskbook/src/components/InjectedComponents/NFT/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './useNFT'
 export * from './useNFTAvatar'
-export * from './useNFTAvatars'
