--- packages/maskbook/src/components/InjectedComponents/NFT/hooks/useNFTAvatars.ts
@@ -1,9 +1,0 @@
-import { useAsyncRetry } from 'react-use'
-import { getNFTAvatars } from '../gun'
-
-export function useNFTAvatars() {
-    return useAsyncRetry(async () => {
-        const result = await getNFTAvatars()
-        return result
-    }, [getNFTAvatars])
-}
